requirements.txt
paho-mqtt>=1.6.1
SQLAlchemy>=2.0.21
pymysql>=1.1.0
python-dotenv>=1.0.0
azure-iot-device>=2.12.0
flask>=3.0.0
Flask-Cors>=4.0.0
tensorflow>=2.14.0
jupyterlab>=4.0.8
pandas>=2.1.3
scikit-learn>=1.3.2
